refactor: clean up finder
